General refactor of folders and private methods
