+++ bias_detector/models/FullNameZipcodeModel.py
